1624 update createalttesterprefab script with the new UI (#1654)
* Modificatio needed for Driver Module

* Modification needed for input Module

* More modification for Input Module

* Modification for CommunicationModule

* Modification for communication module

* Modification needed for UIModule and Logging

* Modification for Editor Module

* More modifications

* Modified editor structure

* Added null checker in inputMisc

* Fixed more errors

* Updated CreateAltPrefab.cs

* Fixed problems with driver

* Added license in coroutine manager

* Update year in inputMisc

* Removed invisible character

* Added file that was deleted

* Added back the WEBGL define

* Refactore code to not use ToUnity method

* Fixed typo

* Fixed prefab problems

* Removed some defines

* Fixed small things in prefab

* Added debug for testing purpose

* Added more logs

* More logs

* More logs

* More logs

* Fix typo in log

* More logs

* fixed component

* More debugs

* Removed Debugs

* Update newInputSystem

* More updates in NewInputSystem

* Update InputController

+++ Assets/AltTester/Runtime/UI/AltDialog.cs
@@ -17,7 +17,10 @@
 
 using System;
 using System.Collections;
+using AltTester.AltTesterUnitySDK.Commands;
 using AltTester.AltTesterUnitySDK.Communication;
+using AltTester.AltTesterUnitySDK.Driver;
+using AltTester.AltTesterUnitySDK.InputModule;
 using AltTester.AltTesterUnitySDK.Logging;
 using UnityEngine;
 #if ENABLE_INPUT_SYSTEM
@@ -174,13 +177,7 @@ namespace AltTester.AltTesterUnitySDK.UI
 
         private void checkIfPlayerPrefNeedsToBeDeleted()
         {
-#if ENABLE_LEGACY_INPUT_MANAGER
-            if (Input.GetKey(KeyCode.LeftControl) && Input.GetKey(KeyCode.LeftShift) && Input.GetKey(KeyCode.D) && Input.GetKey(KeyCode.L))
-#else
-#if ENABLE_INPUT_SYSTEM
-            if (Keyboard.current.leftCtrlKey.isPressed && Keyboard.current.leftShiftKey.isPressed && Keyboard.current.dKey.isPressed && Keyboard.current.lKey.isPressed)
-#endif
-#endif
+            if (InputMisc.IsResetConnectionShortcutPressed())
             {
                 PlayerPrefs.DeleteKey(UID);
                 resetConnectionDataBasedOnUID();
@@ -219,13 +216,9 @@ namespace AltTester.AltTesterUnitySDK.UI
 
         protected IEnumerator SendScreenshot()
         {
-#if UNITY_EDITOR
             if (Application.isBatchMode)
-            {
                 yield return null;
-            }
             else
-#endif
                 yield return new WaitForEndOfFrame();
             this.liveUpdateClient.SendScreenshot();
         }
@@ -378,22 +371,7 @@ namespace AltTester.AltTesterUnitySDK.UI
         {
             CustomInputToggle.isOn = value;
             Icon.color = value ? Color.white : Color.grey;
-
-#if ALTTESTER
-#if ENABLE_LEGACY_INPUT_MANAGER
-            Input.UseCustomInput = value;
-#endif
-#if ENABLE_INPUT_SYSTEM
-            if (value)
-            {
-                NewInputSystem.DisableDefaultDevicesAndEnableAltDevices();
-            }
-            else
-            {
-                NewInputSystem.EnableDefaultDevicesAndDisableAltDevices();
-            }
-#endif
-#endif
+            InputMisc.ActivateCustomInput(value);
         }
 
         private void initCommunicationClient()
